refactor application layer to make it runtime agnostic
